requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
astor==0.8.1
async-timeout==3.0.1
atari-py==0.2.6
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
filelock==3.0.12
funcsigs==1.0.2
future==0.18.2
gast==0.2.2
google==2.0.3
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
gym==0.17.0
h5py==2.10.0
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-rl==0.3.1
keras-rl2==1.0.3
kiwisolver==1.1.0
lz4==3.0.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mock==4.0.1
more-itertools==8.2.0
multidict==4.7.5
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.16.4
oauthlib==3.1.0
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.32
opt-einsum==3.2.0
packaging==20.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.1
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
ray==0.8.2
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
TA-Lib==0.4.17
tabulate==0.8.6
tb-nightly==1.14.0a20190603
tensorboard==1.13.1
tensorboardX==2.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==1.14.0.dev2019060501
Theano==1.0.4
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.0
yarl==1.4.2
zipp==3.1.0
